+++ src/main/scala/org/podval/tools/scalajs/js/LinkTask.scala
@@ -20,14 +20,25 @@ sealed abstract class LinkTask extends DefaultTask with ScalaJSTask:
 
   final override protected def linkTask: LinkTask = this
 
-  def moduleInitializerProperties: Option[Seq[ModuleInitializerProperties]]
+  def moduleInitializers: Option[Seq[ModuleInitializer]]
 
   @InputFiles def getRuntimeClassPath: ConfigurableFileCollection
-  @Input @Optional def getModuleKind      : Property[String]
+  
+  @Input @Optional def getPrettyPrint: Property[Boolean]
+  final def prettyPrint: Boolean = 
+    getPrettyPrint.getOrElse(false)
+    
+  @Input @Optional def getModuleKind: Property[String]
+  final def moduleKind: ModuleKind = 
+    LinkTask.byName(getModuleKind, ModuleKind.NoModule, ModuleKind.values.toList)
+  
   @Input @Optional def getModuleSplitStyle: Property[String]
-  @Input @Optional def getPrettyPrint     : Property[Boolean]
-  @Input @Optional def getOptimization    : Property[String]
-  final def optimization: Optimization = LinkTask.byName(getOptimization, Optimization.Fast, Optimization.values.toList)
+  final def moduleSplitStyle: ModuleSplitStyle = 
+    LinkTask.byName(getModuleSplitStyle, ModuleSplitStyle.FewestModules, ModuleSplitStyle.values.toList)
+  
+  @Input @Optional def getOptimization: Property[String]
+  final def optimization: Optimization = 
+    LinkTask.byName(getOptimization, Optimization.Fast, Optimization.values.toList)
 
   private val buildDirectory: File = getProject.getLayout.getBuildDirectory.get.getAsFile
   private def outputFile(name: String): File = Files.file(buildDirectory, "scalajs", getName, name)
@@ -36,19 +47,18 @@ sealed abstract class LinkTask extends DefaultTask with ScalaJSTask:
   @OutputFile final def getReportBinFile : File = outputFile("linking-report.bin")
 
 object LinkTask:
-  def byName[T](property: Property[String], default: => T, all: List[T]): T =
+  private def byName[T](property: Property[String], default: => T, all: List[T]): T =
     if !property.isPresent then default else all.find(_.toString == property.get).get
 
   abstract class Main extends LinkTask:
     final override protected def flavour: String = "Link"
     final override def sourceSetName: String = SourceSet.MAIN_SOURCE_SET_NAME
     @Nested def getModuleInitializers: NamedDomainObjectContainer[ModuleInitializerProperties]
-    final override def moduleInitializerProperties: Option[Seq[ModuleInitializerProperties]] =
-      Some(getModuleInitializers.asScala.toSeq)
+    final override def moduleInitializers: Option[Seq[ModuleInitializer]] =
+      Some(getModuleInitializers.asScala.toSeq.map(_.toModuleInitializer))
 
   abstract class Test extends LinkTask:
     final override protected def flavour: String = "LinkTest"
     final override def sourceSetName: String = SourceSet.TEST_SOURCE_SET_NAME
-    final override def moduleInitializerProperties: Option[Seq[ModuleInitializerProperties]] =
+    final override def moduleInitializers: Option[Seq[ModuleInitializer]] =
       None
-