+++ src/main/scala/org/podval/tools/scalajs/js/ModuleKind.scala
@@ -1,0 +1,6 @@
+package org.podval.tools.scalajs.js
+
+enum ModuleKind derives CanEqual:
+  case NoModule
+  case ESModule
+  case CommonJSModule