+++ src/main/scala/org/podval/tools/scalajs/js/ModuleSplitStyle.scala
@@ -1,0 +1,5 @@
+package org.podval.tools.scalajs.js
+
+enum ModuleSplitStyle derives CanEqual:
+  case FewestModules
+  case SmallestModules